Moved all_nodes_survey_graph to a separate file.
